refactor: move update dialog to overlays and convert to overlay workflow
