+++ apps/flites/lib/screens/overview.dart
@@ -3,10 +3,10 @@ import 'package:flites/services/update_service.dart';
 import 'package:flites/types/update_info.dart';
 import 'package:flites/widgets/blocking_widget/blocking_container.dart';
 import 'package:flites/widgets/image_editor/image_editor.dart';
+import 'package:flites/widgets/overlays/update_overlay.dart';
 import 'package:flites/widgets/project_file_list/project_file_list_vertical.dart';
 import 'package:flites/widgets/tool_controls/tool_controls.dart';
 import 'package:flites/widgets/tool_controls/zoom_controls.dart';
-import 'package:flites/widgets/update/update_dialog.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
@@ -37,28 +37,8 @@ class _OverviewState extends State<Overview> {
     final bool hasUpdate = updateInfo != null;
 
     if (hasUpdate && mounted) {
-      final shouldUpdate = await showDialog<bool>(
-            context: context,
-            barrierDismissible: false,
-            barrierColor: Colors.transparent,
-            builder: (_) => UpdateDialog(
-              currentVersion: updateInfo.currentVersion,
-              newVersion: updateInfo.newVersion,
-            ),
-          ) ??
-          false;
-
-      if (shouldUpdate && mounted) {
-        final success = await UpdateService.getReleaseAndLaunchUrl();
-        if (!success && mounted) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content:
-                  Text('Failed to perform update. Please try again later.'),
-            ),
-          );
-        }
-      }
+      showUpdateOverlay.value = true;
+      updateOverlayInfo.value = updateInfo;
     }
   }
